rvps: consolidate native.rs into lib.rs
As is, you might guess that native.rs would hold code that is specific
to the built-in RVPS, but it doesn't. Instead it holds the main
top-level logic of the RVPS. Following other components in this repo,
let's move this to lib.rs.
