+++ rvps/src/lib.rs
@@ -10,14 +10,16 @@ pub mod reference_value;
 pub mod storage;
 
 pub use config::Config;
-
-pub mod native;
-pub use native::Core;
-
-use serde::{Deserialize, Serialize};
-
 pub use reference_value::{ReferenceValue, TrustedDigest};
 pub use storage::ReferenceValueStorage;
+
+use extractors::{Extractors, ExtractorsImpl};
+use pre_processor::{PreProcessor, PreProcessorAPI};
+
+use anyhow::{bail, Context, Result};
+use log::{info, warn};
+use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
 
 /// Default version of Message
 static MESSAGE_VERSION: &str = "0.1.0";
@@ -41,4 +43,78 @@ pub struct Message {
 /// Set the default version for Message
 fn default_version() -> String {
     MESSAGE_VERSION.into()
+}
+
+/// The core of the RVPS, s.t. componants except communication componants.
+pub struct Core {
+    pre_processor: PreProcessor,
+    extractors: ExtractorsImpl,
+    storage: Box<dyn ReferenceValueStorage + Send + Sync>,
+}
+
+impl Core {
+    /// Instantiate  a new RVPS Core
+    pub fn new(config: Config) -> Result<Self> {
+        let pre_processor = PreProcessor::default();
+        let extractors = ExtractorsImpl::default();
+        let storage = config.storage.to_storage()?;
+
+        Ok(Core {
+            pre_processor,
+            extractors,
+            storage,
+        })
+    }
+
+    /// Add Ware to the Core's Pre-Processor
+    pub fn with_ware(&mut self, _ware: &str) -> &Self {
+        // TODO: no wares implemented now.
+        self
+    }
+
+    pub async fn verify_and_extract(&mut self, message: &str) -> Result<()> {
+        let mut message: Message = serde_json::from_str(message).context("parse message")?;
+
+        // Judge the version field
+        if message.version != MESSAGE_VERSION {
+            bail!(
+                "Version unmatched! Need {}, given {}.",
+                MESSAGE_VERSION,
+                message.version
+            );
+        }
+
+        self.pre_processor.process(&mut message)?;
+
+        let rv = self.extractors.process(message)?;
+        for v in rv.iter() {
+            let old = self.storage.set(v.name().to_string(), v.clone()).await?;
+            if let Some(old) = old {
+                info!("Old Reference value of {} is replaced.", old.name());
+            }
+        }
+
+        Ok(())
+    }
+
+    pub async fn get_digests(&self) -> Result<HashMap<String, Vec<String>>> {
+        let mut rv_map = HashMap::new();
+        let reference_values = self.storage.get_values().await?;
+
+        for rv in reference_values {
+            if rv.expired() {
+                warn!("Reference value of {} is expired.", rv.name());
+                continue;
+            }
+
+            let hash_values = rv
+                .hash_values()
+                .iter()
+                .map(|pair| pair.value().to_owned())
+                .collect();
+
+            rv_map.insert(rv.name().to_string(), hash_values);
+        }
+        Ok(rv_map)
+    }
 }
